+++ src/EncloseField.php
@@ -37,17 +37,13 @@ class EncloseField extends php_user_filter
 
     /**
      * Default sequence.
-     *
-     * @var string
      */
-    protected $sequence;
+    protected string $sequence;
 
     /**
      * Characters that triggers enclosure in PHP.
-     *
-     * @var string
      */
-    protected static $force_enclosure = "\n\r\t ";
+    protected static string $force_enclosure = "\n\r\t ";
 
     /**
      * Static method to return the stream filter filtername.
@@ -81,14 +77,8 @@ class EncloseField extends php_user_filter
             throw new InvalidArgumentException('The sequence must contain at least one character to force enclosure');
         }
 
-        $formatter = function (array $record) use ($sequence): array {
-            return array_map(function (?string $value) use ($sequence): string {
-                return $sequence.$value;
-            }, $record);
-        };
-
         return $csv
-            ->addFormatter($formatter)
+            ->addFormatter(fn (array $record): array => array_map(fn (?string $value): string => $sequence.$value, $record))
             ->addStreamFilter(self::FILTERNAME, ['sequence' => $sequence]);
     }
 
@@ -119,10 +109,10 @@ class EncloseField extends php_user_filter
      */
     public function filter($in, $out, &$consumed, $closing): int
     {
-        while ($res = stream_bucket_make_writeable($in)) {
-            $res->data = str_replace($this->params['sequence'], '', $res->data);
-            $consumed += $res->datalen;
-            stream_bucket_append($out, $res);
+        while (null !== ($bucket = stream_bucket_make_writeable($in))) {
+            $bucket->data = str_replace($this->params['sequence'], '', $bucket->data);
+            $consumed += $bucket->datalen;
+            stream_bucket_append($out, $bucket);
         }
 
         return PSFS_PASS_ON;