+++ src/RFC4180Field.php
@@ -47,21 +47,20 @@ class RFC4180Field extends php_user_filter
      *
      * @var string[]
      */
-    protected $search;
+    protected array $search;
 
     /**
      * The replacement value that replace found $search values.
      *
      * @var string[]
      */
-    protected $replace;
+    protected array $replace;
 
     /**
      * Characters that triggers enclosure with PHP fputcsv.
      *
-     * @var string
      */
-    protected static $force_enclosure = "\n\r\t ";
+    protected static string $force_enclosure = "\n\r\t ";
 
     /**
      * Static method to add the stream filter to a {@link AbstractCsv} object.
@@ -102,11 +101,7 @@ class RFC4180Field extends php_user_filter
             return $value;
         };
 
-        $formatter = static function (array $record) use ($mapper): array {
-            return array_map($mapper, $record);
-        };
-
-        return $csv->addFormatter($formatter);
+        return $csv->addFormatter(fn (array $record): array => array_map($mapper, $record));
     }
 
     /**
@@ -135,7 +130,7 @@ class RFC4180Field extends php_user_filter
      */
     public function filter($in, $out, &$consumed, $closing): int
     {
-        while ($bucket = stream_bucket_make_writeable($in)) {
+        while (null !== ($bucket = stream_bucket_make_writeable($in))) {
             $bucket->data = str_replace($this->search, $this->replace, $bucket->data);
             $consumed += $bucket->datalen;
             stream_bucket_append($out, $bucket);