+++ src/ResultSetTest.php
@@ -26,20 +26,9 @@ use function next;
  */
 final class ResultSetTest extends TestCase
 {
-    /**
-     * @var Reader
-     */
-    private $csv;
-
-    /**
-     * @var Statement
-     */
-    private $stmt;
-
-    /**
-     * @var array
-     */
-    private $expected = [
+    private Reader $csv;
+    private Statement $stmt;
+    private array $expected = [
         ['john', 'doe', 'john.doe@example.com'],
         ['jane', 'doe', 'jane.doe@example.com'],
     ];