+++ src/XMLConverter.php
@@ -25,38 +25,28 @@ class XMLConverter
 {
     /**
      * XML Root name.
-     *
-     * @var string
      */
-    protected $root_name = 'csv';
+    protected string $root_name = 'csv';
 
     /**
      * XML Node name.
-     *
-     * @var string
      */
-    protected $record_name = 'row';
+    protected string $record_name = 'row';
 
     /**
      * XML Item name.
-     *
-     * @var string
      */
-    protected $field_name = 'cell';
+    protected string $field_name = 'cell';
 
     /**
      * XML column attribute name.
-     *
-     * @var string
      */
-    protected $column_attr = '';
+    protected string $column_attr = '';
 
     /**
      * XML offset attribute name.
-     *
-     * @var string
      */
-    protected $offset_attr = '';
+    protected string $offset_attr = '';
 
     public static function create(): self
     {