+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All Notable changes to `Csv` will be documented in this file
 
 ### Added
 
-- None
+- Added PHP7.4 typed properties where applicable
 
 ### Deprecated
 
@@ -14,11 +14,14 @@ All Notable changes to `Csv` will be documented in this file
 
 ### Fixed
 
-- None
+- `AbstractCsv` constructor is marked final via docblock.  
+ The method should never be extended or changed in child classes to avoid unexpected behaviour
 
 ### Removed
 
-- None
+- PHP7.3 support
+- Remove internal `EmptyEscapeParser` Polyfill used in `Reader` class
+- Remove PHP7.4 polyfill features in `Writer` class
 
 ## 9.7.4 - 2021-11-30